Minor changes in addition to previous commit

+++ Windows/Timer.cpp
@@ -17,7 +17,7 @@ void Timer::Start()
 	_sampleCounter++;
 }
 
-void Timer::Stop(int numBytesMoved)
+void Timer::Stop(size_t numBytesMoved)
 {
 	_accumulatedBytes += numBytesMoved;
